Revert "Fixed compile errors"
This reverts commit 46acfbe27114b79ee51e4c42a74cb738b18c9c2e.

+++ contracts/UniProxy.sol
@@ -221,7 +221,7 @@ contract UniProxy is ReentrancyGuard {
 
       /// tick(imprecise as it's an integer) to price
       sqrtPriceX96 = TickMath.getSqrtRatioAtTick(
-        int24((tickCumulatives[1] - tickCumulatives[0]) / int256(uint256(_twapInterval)))
+        int24((tickCumulatives[1] - tickCumulatives[0]) / _twapInterval)
       );
     }
   }
